Hopefully make travis pass.
My tests need the module `responses` to run, as I use it to mock
`requests`. However, I changed things around so that the tests package
is not installed, and `responses` is not a dependency. It is however a
dependendency for testing in `setup.py`, so running
`python setup.py test` will install responses specifically for the tests.

My current solution is to make the Dockerfile install responses, and
hopefully that will work. I wonder if it is a bit of a hack though, and
if we should use `python setup.py test` as the testing mechanism.
I'm not sure!